refactor(147365): renomeia arquivo de recreio_cei para recreio_cei_cci_cips
